+++ sound150@claudiux/files/sound150@claudiux/6.4/lib/s150PopupMenu.js
@@ -294,7 +294,19 @@ class Player extends PopupMenu.PopupMenuSection {
         });
         this.artistLabel.clutterText.line_wrap = true;
         this.artistLabel.clutterText.line_wrap_mode = Pango.WrapMode.WORD_CHAR;
-        this.artistLabel.clutterText.ellipsize = Pango.EllipsizeMode.NONE;
+        switch (this._applet.shortenArtistTitle) {
+            case "NONE":
+                this.artistLabel.clutterText.ellipsize = Pango.EllipsizeMode.NONE;
+                break;
+            case "START":
+                this.artistLabel.clutterText.ellipsize = Pango.EllipsizeMode.START;
+                break;
+            case "MIDDLE":
+                this.artistLabel.clutterText.ellipsize = Pango.EllipsizeMode.MIDDLE;
+                break;
+            case "END":
+                this.artistLabel.clutterText.ellipsize = Pango.EllipsizeMode.END;
+        }
         artistInfo.add_actor(artistIcon);
         artistInfo.add_actor(this.artistLabel);
         let titleInfo = new St.BoxLayout();
@@ -309,7 +321,19 @@ class Player extends PopupMenu.PopupMenuSection {
 
         this.titleLabel.clutterText.line_wrap = true;
         this.titleLabel.clutterText.line_wrap_mode = Pango.WrapMode.WORD_CHAR;
-        this.titleLabel.clutterText.ellipsize = Pango.EllipsizeMode.NONE;
+        switch (this._applet.shortenArtistTitle) {
+            case "NONE":
+                this.titleLabel.clutterText.ellipsize = Pango.EllipsizeMode.NONE;
+                break;
+            case "START":
+                this.titleLabel.clutterText.ellipsize = Pango.EllipsizeMode.START;
+                break;
+            case "MIDDLE":
+                this.titleLabel.clutterText.ellipsize = Pango.EllipsizeMode.MIDDLE;
+                break;
+            case "END":
+                this.titleLabel.clutterText.ellipsize = Pango.EllipsizeMode.END;
+        }
         titleInfo.add_actor(titleIcon);
         titleInfo.add_actor(this.titleLabel);
         //~ titleInfo.add_actor(this.display_cover_button.getActor());
@@ -860,16 +884,16 @@ class Player extends PopupMenu.PopupMenuSection {
             this._cover_path = cover_path;
             this._applet._icon_path = cover_path; // Added
             this._applet.setAppletIcon(this._applet.player, cover_path); // Added
-            this._cover_load_handle = St.TextureCache.get_default().load_image_from_file_async(
-                cover_path,
-                Math.trunc(300 * this._applet.real_ui_scale),
-                Math.trunc(300 * this._applet.real_ui_scale),
-                (cache, handle, actor) => {
-                    this._on_cover_loaded(cache, handle, actor)
-                }
-            );
+            if (cover_path != null)
+                this._cover_load_handle = St.TextureCache.get_default().load_image_from_file_async(
+                    cover_path,
+                    Math.trunc(300 * this._applet.real_ui_scale),
+                    Math.trunc(300 * this._applet.real_ui_scale),
+                    (cache, handle, actor) => {
+                        this._on_cover_loaded(cache, handle, actor)
+                    }
+                );
             this._applet.setIcon();
-
 
             //~ log("this._cover_path: "+this._cover_path, true);
             try {
@@ -920,13 +944,13 @@ class Player extends PopupMenu.PopupMenuSection {
             return;
         }
 
-        try {
-            if (this.coverBox != null && this.cover != null) {
-                let coverBoxChildren = this.coverBox.get_children();
-                if (coverBoxChildren.length > 0 && coverBoxChildren.indexOf(this.cover) > -1)
-                    this.coverBox.remove_child(this.cover);
-            }
-        } catch (e) {}
+        //~ try {
+        if (this.coverBox != null && this.cover != null) {
+            let coverBoxChildren = this.coverBox.get_children();
+            if (coverBoxChildren.length > 0 && coverBoxChildren.indexOf(this.cover) > -1)
+                try { this.coverBox.remove_child(this.cover) } catch(e) {}
+        }
+        //~ } catch (e) {}
 
         // Make sure any oddly-shaped album art doesn't affect the height of the applet popup
         // (and move the player controls as a result).
@@ -968,9 +992,13 @@ class Player extends PopupMenu.PopupMenuSection {
                     this.coverBox.add_actor(this.cover);
                     //~ this.coverBox.set_reactive = true;
                     //~ this.coverBox.connect("button-press-event", (event) => Util.spawnCommandLineAsync("xdg-open "+this._cover_path));
-                    this.coverBox.set_child_below_sibling(this.cover, this.trackInfo);
+                    try {
+                        this.coverBox.set_child_below_sibling(this.cover, this.trackInfo);
+                    } catch(e) {}
                 } else {
-                    this.coverBox.set_child_below_sibling(this.trackInfo, null);
+                    try {
+                        this.coverBox.set_child_below_sibling(this.trackInfo, null);
+                    } catch(e) {}
                 }
             }
         } catch (e) {}