+++ Direct@claudiux/files/Direct@claudiux/CHANGELOG.md
@@ -1,3 +1,9 @@
+### v1.3.1~20251115
+* Adds option "Open the menu hovering over the icon".
+
+### v1.3.0~20251112
+* Adds the Favorite Apps section.
+
 ### v1.2.0~20251109
 * Anticipates likely future configurations for Cinnamon.